requirements.txt
fastapi
uvicorn
openai
python-dotenv
mmh3
requests
jinja2
pydantic[email]
python-multipart
httpx
pymilvus
boto3
botocore
supabase
python-jose[cryptography]
passlib[bcrypt]
itsdangerous